Organised in following packages: config, implementation and mediation. Tested fixed for renamed /command
