+++ src/main/java/jacamo/rest/mediation/TranslOrg.java
@@ -1,4 +1,4 @@
-package jacamo.rest;
+package jacamo.rest.mediation;
 
 import java.util.ArrayList;
 import java.util.HashMap;
